+++ mitreattack/diffStix/changelog_helper.py
@@ -1,5 +1,6 @@
 import argparse
 import datetime
+import math
 import json
 import os
 from itertools import chain
@@ -20,13 +21,14 @@ domainToDomainLabel = {"enterprise-attack": "Enterprise", "mobile-attack": "Mobi
 domainToTaxiiCollectionId = {
     "enterprise-attack": "95ecc380-afe9-11e4-9b6c-751b66dd541e",
     "mobile-attack": "2f669986-b40b-4423-b720-4396ca6a462b",
-    "ics-attack": "02c3ef24-9cd4-48f3-a99f-b74ce24f1d34"
+    "ics-attack": "02c3ef24-9cd4-48f3-a99f-b74ce24f1d34",
 }
 # stix filters for querying for each type of data
 attackTypeToStixFilter = {
     "technique": [Filter("type", "=", "attack-pattern")],
     "software": [Filter("type", "=", "malware"), Filter("type", "=", "tool")],
     "group": [Filter("type", "=", "intrusion-set")],
+    "campaign": [Filter("type", "=", "campaign")],
     "mitigation": [Filter("type", "=", "course-of-action")],
     "datasource": [
         Filter("type", "=", "x-mitre-data-source"),
@@ -40,6 +42,7 @@ attackTypeToTitle = {
     "malware": "Malware",
     "software": "Software",
     "group": "Groups",
+    "campaign": "Campaigns",
     "mitigation": "Mitigations",
     "datasource": "Data Sources and/or Components",
 }
@@ -49,17 +52,18 @@ attackTypeToSectionName = {
     "malware": "Malware",
     "software": "Software",
     "group": "Group",
+    "campaign": "Campaign",
     "mitigation": "Mitigation",
     "datasource": "Data Source and/or Component",
 }
 # how we want to format headers for each section
 sectionNameToSectionHeaders = {
     "additions": "New {obj_type}",
-    "changes": "{obj_type} changes",
-    "minor_changes": "Minor {obj_type} changes",
-    "deprecations": "{obj_type} deprecations",
-    "revocations": "{obj_type} revocations",
-    "deletions": "{obj_type} deletions",
+    "changes": "{obj_type} Changes",
+    "minor_changes": "Minor {obj_type} Changes",
+    "deprecations": "{obj_type} Deprecations",
+    "revocations": "{obj_type} Revocations",
+    "deletions": "{obj_type} Deletions",
     "unchanged": "Unchanged {obj_type}",
 }
 # color key for layers
@@ -93,6 +97,7 @@ layer_defaults = [
 md_default = os.path.join("output", f"updates-{this_month.lower()}.md")
 json_default = os.path.join("output", f"updates-{this_month.lower()}.json")
 
+
 class DiffStix(object):
     """Utilities for detecting and summarizing differences between two versions of the ATT&CK content."""
 
@@ -107,7 +112,7 @@ class DiffStix(object):
         old="old",
         show_key=False,
         site_prefix="",
-        types=["technique", "software", "group", "mitigation", "datasource"],
+        types=["technique", "software", "group", "campaign", "mitigation", "datasource"],
         use_taxii=False,
         use_mitre_cti=False,
         verbose=False,
@@ -188,26 +193,18 @@ class DiffStix(object):
 
         logger.info("removing duplicate relationships")
         self.new_subtechnique_of_rels = [
-            i
-            for n, i in enumerate(self.new_subtechnique_of_rels)
-            if i not in self.new_subtechnique_of_rels[n + 1 :]
+            i for n, i in enumerate(self.new_subtechnique_of_rels) if i not in self.new_subtechnique_of_rels[n + 1 :]
         ]
         self.old_subtechnique_of_rels = [
-            i
-            for n, i in enumerate(self.old_subtechnique_of_rels)
-            if i not in self.old_subtechnique_of_rels[n + 1 :]
+            i for n, i in enumerate(self.old_subtechnique_of_rels) if i not in self.old_subtechnique_of_rels[n + 1 :]
         ]
 
         logger.info("removing duplicate data components")
         self.new_datacomponents = [
-            i
-            for n, i in enumerate(self.new_datacomponents)
-            if i not in self.new_datacomponents[n + 1 :]
+            i for n, i in enumerate(self.new_datacomponents) if i not in self.new_datacomponents[n + 1 :]
         ]
         self.old_datacomponents = [
-            i
-            for n, i in enumerate(self.old_datacomponents)
-            if i not in self.old_datacomponents[n + 1 :]
+            i for n, i in enumerate(self.old_datacomponents) if i not in self.old_datacomponents[n + 1 :]
         ]
 
     def getUrlFromStix(self, datum, is_subtechnique=False):
@@ -230,6 +227,8 @@ class DiffStix(object):
         """Transform stix to dict and deep copy the dict."""
         result = []
         for obj in objects:
+            # TODO: serialize the STIX objects instead of casting them to dict
+            # more details here: https://github.com/mitre/cti/issues/17#issuecomment-395768815
             obj = dict(obj)
             if "external_references" in obj:
                 for i in range(len(obj["external_references"])):
@@ -237,8 +236,14 @@ class DiffStix(object):
             if "kill_chain_phases" in obj:
                 for i in range(len(obj["kill_chain_phases"])):
                     obj["kill_chain_phases"][i] = dict(obj["kill_chain_phases"][i])
+
             if "modified" in obj:
                 obj["modified"] = str(obj["modified"])
+            if "first_seen" in obj:
+                obj["first_seen"] = str(obj["first_seen"])
+            if "last_seen" in obj:
+                obj["last_seen"] = str(obj["last_seen"])
+
             if "definition" in obj:
                 obj["definition"] = dict(obj["definition"])
             obj["created"] = str(obj["created"])
@@ -258,12 +263,7 @@ class DiffStix(object):
 
                 def load_datastore(data_store):
                     """Handle data loaded from either a directory or the TAXII server"""
-                    raw_data = list(
-                        chain.from_iterable(
-                            data_store.query(f)
-                            for f in attackTypeToStixFilter[obj_type]
-                        )
-                    )
+                    raw_data = list(chain.from_iterable(data_store.query(f) for f in attackTypeToStixFilter[obj_type]))
                     raw_data = self.deep_copy_stix(raw_data)
                     id_to_obj = {item["id"]: item for item in raw_data}
 
@@ -276,9 +276,7 @@ class DiffStix(object):
                 def parse_subtechniques(data_store, new=False):
                     """Parse dataStore sub-technique-of relationships"""
                     if new:
-                        for technique in list(
-                            data_store.query(attackTypeToStixFilter["technique"])
-                        ):
+                        for technique in list(data_store.query(attackTypeToStixFilter["technique"])):
                             self.new_id_to_technique[technique["id"]] = technique
                         self.new_subtechnique_of_rels += list(
                             data_store.query(
@@ -289,9 +287,7 @@ class DiffStix(object):
                             )
                         )
                     else:
-                        for technique in list(
-                            data_store.query(attackTypeToStixFilter["technique"])
-                        ):
+                        for technique in list(data_store.query(attackTypeToStixFilter["technique"])):
                             self.old_id_to_technique[technique["id"]] = technique
                         self.old_subtechnique_of_rels += list(
                             data_store.query(
@@ -305,48 +301,32 @@ class DiffStix(object):
                 def parse_datacomponents(data_store, new=False):
                     """Parse dataStore x-mitre-data-components"""
                     if new:
-                        for datasource in list(
-                            data_store.query(attackTypeToStixFilter["datasource-only"])
-                        ):
+                        for datasource in list(data_store.query(attackTypeToStixFilter["datasource-only"])):
                             self.new_id_to_datasource[datasource["id"]] = datasource
                         self.new_datacomponents += list(
-                            data_store.query(
-                                [Filter("type", "=", "x-mitre-data-component")]
-                            )
+                            data_store.query([Filter("type", "=", "x-mitre-data-component")])
                         )
                     else:
-                        for datasource in list(
-                            data_store.query(attackTypeToStixFilter["datasource-only"])
-                        ):
+                        for datasource in list(data_store.query(attackTypeToStixFilter["datasource-only"])):
                             self.old_id_to_datasource[datasource["id"]] = datasource
                         self.old_datacomponents += list(
-                            data_store.query(
-                                [Filter("type", "=", "x-mitre-data-component")]
-                            )
+                            data_store.query([Filter("type", "=", "x-mitre-data-component")])
                         )
 
                 def update_contributors(old_object, new_object):
                     """Update contributors list if new object has contributors"""
                     if new_object.get("x_mitre_contributors"):
-                        new_object_contributors = set(
-                            new_object["x_mitre_contributors"]
-                        )
+                        new_object_contributors = set(new_object["x_mitre_contributors"])
 
                         # Check if old objects had contributors
-                        if old_object is None or not old_object.get(
-                            "x_mitre_contributors"
-                        ):
+                        if old_object is None or not old_object.get("x_mitre_contributors"):
                             old_object_contributors = set()
                         else:
-                            old_object_contributors = set(
-                                old_object["x_mitre_contributors"]
-                            )
+                            old_object_contributors = set(old_object["x_mitre_contributors"])
 
                         # Remove old contributors from showing up
                         # if contributors are the same the result will be empty
-                        new_contributors = (
-                            new_object_contributors - old_object_contributors
-                        )
+                        new_contributors = new_object_contributors - old_object_contributors
 
                         # Update counter of contributor to track contributions
                         for new_contributor in new_contributors:
@@ -367,8 +347,7 @@ class DiffStix(object):
                 def load_taxii(new=False):
                     """Load data from TAXII server according to domain"""
                     collection = Collection(
-                        "https://cti-taxii.mitre.org/stix/collections/"
-                        + domainToTaxiiCollectionId[domain]
+                        "https://cti-taxii.mitre.org/stix/collections/" + domainToTaxiiCollectionId[domain]
                     )
                     data_store = TAXIICollectionSource(collection)
                     parse_subtechniques(data_store, new)
@@ -412,101 +391,100 @@ class DiffStix(object):
                 unchanged = set()
 
                 # find changes, revocations and deprecations
-                for key in intersection:
+                for stix_id in intersection:
+
+                    old_stix_obj = old["id_to_obj"][stix_id]
+                    new_stix_obj = new["id_to_obj"][stix_id]
+                    if new_stix_obj["type"] == "x-mitre-data-component":
+                        attack_id = None
+                    else:
+                        attack_id = new_stix_obj["external_references"][0]["external_id"]
 
                     # find revoked objects in the new bundle
-                    if (
-                        "revoked" in new["id_to_obj"][key]
-                        and new["id_to_obj"][key]["revoked"]
-                    ):
+                    if "revoked" in new_stix_obj and new_stix_obj["revoked"]:
                         # only work with newly revoked objects
-                        if (
-                            not "revoked" in old["id_to_obj"][key]
-                            or not old["id_to_obj"][key]["revoked"]
-                        ):
+                        if not "revoked" in old_stix_obj or not old_stix_obj["revoked"]:
                             # store the revoking object
                             revoked_by_key = new["data_store"].query(
                                 [
                                     Filter("type", "=", "relationship"),
                                     Filter("relationship_type", "=", "revoked-by"),
-                                    Filter("source_ref", "=", key),
+                                    Filter("source_ref", "=", stix_id),
                                 ]
                             )
                             if len(revoked_by_key) == 0:
-                                logger.error(
-                                    f"[{key}] revoked object has no revoked-by relationship"
-                                )
+                                logger.error(f"[{stix_id}] revoked object has no revoked-by relationship")
                                 continue
                             else:
                                 revoked_by_key = revoked_by_key[0]["target_ref"]
 
                             if revoked_by_key not in new["id_to_obj"]:
-                                logger.error(f"{key} revoked by {revoked_by_key}, but {revoked_by_key} not found in new STIX bundle!!")
+                                logger.error(
+                                    f"{stix_id} revoked by {revoked_by_key}, but {revoked_by_key} not found in new STIX bundle!!"
+                                )
                                 continue
 
-                            new["id_to_obj"][key]["revoked_by"] = new["id_to_obj"][
-                                revoked_by_key
-                            ]
+                            new_stix_obj["revoked_by"] = new["id_to_obj"][revoked_by_key]
 
-                            revocations.add(key)
+                            revocations.add(stix_id)
                         # else it was already revoked, and not a change; do nothing with it
 
                     # find deprecated objects
-                    elif (
-                        "x_mitre_deprecated" in new["id_to_obj"][key]
-                        and new["id_to_obj"][key]["x_mitre_deprecated"]
-                    ):
+                    elif "x_mitre_deprecated" in new_stix_obj and new_stix_obj["x_mitre_deprecated"]:
                         # if previously deprecated, not a change
-                        if not "x_mitre_deprecated" in old["id_to_obj"][key]:
-                            deprecations.add(key)
+                        if not "x_mitre_deprecated" in old_stix_obj:
+                            deprecations.add(stix_id)
 
                     # find all other changed objects
                     else:
                         # try getting version numbers; should only lack version numbers if something has gone
                         # horribly wrong or a revoked object has slipped through
                         try:
-                            old_version = float(
-                                old["id_to_obj"][key]["x_mitre_version"]
-                            )
+                            old_version = float(old_stix_obj["x_mitre_version"])
                         except ValueError:
-                            logger.error(
-                                f"ERROR: cannot get old version for object: {key}"
-                            )
+                            logger.error(f"ERROR: cannot get old version for object: {stix_id}")
 
                         try:
-                            new_version = float(
-                                new["id_to_obj"][key]["x_mitre_version"]
-                            )
+                            new_version = float(new_stix_obj["x_mitre_version"])
                         except ValueError:
-                            logger.error(
-                                f"ERROR: cannot get new version for object: {key}"
-                            )
+                            logger.error(f"ERROR: cannot get new version for object: {stix_id}")
 
                         # Verify if there are new contributors on the object
-                        update_contributors(
-                            old["id_to_obj"][key], new["id_to_obj"][key]
-                        )
+                        update_contributors(old_stix_obj, new_stix_obj)
 
-                        # check for changes
+                        # check for changes, store previous version for display
+                        new_stix_obj["previous_version"] = old_version
                         if new_version > old_version:
                             # an update has occurred to this object
-                            changes.add(key)
+                            changes.add(stix_id)
+
+                            # verify version change
+                            if not self.version_increment_is_valid(str(old_version), str(new_version), "changes"):
+                                logger.warning(
+                                    f"{stix_id} - Unexpected version. [{attack_id}] {new_stix_obj['name']} incremented from {old_version} to {new_version}"
+                                )
                         else:
                             # check for minor change; modification date increased but not version
-                            old_date = dateparser.parse(
-                                old["id_to_obj"][key]["modified"]
-                            )
-                            new_date = dateparser.parse(
-                                new["id_to_obj"][key]["modified"]
-                            )
+                            old_date = dateparser.parse(old_stix_obj["modified"])
+                            new_date = dateparser.parse(new_stix_obj["modified"])
                             if new_date > old_date:
-                                minor_changes.add(key)
+                                minor_changes.add(stix_id)
                             else:
-                                unchanged.add(key)
+                                unchanged.add(stix_id)
 
                 # Add contributions from additions
-                for key in additions:
-                    update_contributors(None, new["id_to_obj"][key])
+                for stix_id in additions:
+                    new_stix_obj = new["id_to_obj"][stix_id]
+                    if new_stix_obj["type"] == "x-mitre-data-component":
+                        attack_id = None
+                    else:
+                        attack_id = new_stix_obj["external_references"][0]["external_id"]
+                    update_contributors(None, new_stix_obj)
+                    # verify version is 1.0
+                    if not self.version_increment_is_valid(None, new_stix_obj["x_mitre_version"], "additions"):
+                        logger.warning(
+                            f"{stix_id} - Unexpected version. [{attack_id}] {new_stix_obj['name']} is new and has version {new_stix_obj['x_mitre_version']}. Expected 1.0"
+                        )
 
                 # set data
                 if obj_type not in self.data:
@@ -517,32 +495,42 @@ class DiffStix(object):
                 }
                 # only create minor_changes data if we want to display it later
                 if self.minor_changes:
-                    self.data[obj_type][domain]["minor_changes"] = [
-                        new["id_to_obj"][key] for key in minor_changes
-                    ]
+                    self.data[obj_type][domain]["minor_changes"] = [new["id_to_obj"][key] for key in minor_changes]
 
                 # ditto for unchanged
                 if self.unchanged:
-                    self.data[obj_type][domain]["unchanged"] = [
-                        new["id_to_obj"][key] for key in unchanged
-                    ]
+                    self.data[obj_type][domain]["unchanged"] = [new["id_to_obj"][key] for key in unchanged]
 
-                self.data[obj_type][domain]["revocations"] = [
-                    new["id_to_obj"][key] for key in revocations
-                ]
-                self.data[obj_type][domain]["deprecations"] = [
-                    new["id_to_obj"][key] for key in deprecations
-                ]
+                self.data[obj_type][domain]["revocations"] = [new["id_to_obj"][key] for key in revocations]
+                self.data[obj_type][domain]["deprecations"] = [new["id_to_obj"][key] for key in deprecations]
 
                 # only show deletions if objects were deleted
                 if len(deletions) > 0:
-                    self.data[obj_type][domain]["deletions"] = [
-                        old["id_to_obj"][key] for key in deletions
-                    ]
+                    self.data[obj_type][domain]["deletions"] = [old["id_to_obj"][key] for key in deletions]
 
                 logger.debug(f"Loaded:  [{domain:17}]/{obj_type}")
                 pbar.update(1)
         pbar.close()
+
+    def version_increment_is_valid(self, old_version: str, new_version: str, section: str) -> bool:
+        """Validate version increment between old and new STIX objects."""
+        if section in ["revocations", "deprecations"]:
+            return True
+        if section == "additions":
+            if new_version != "1.0":
+                return False
+            return True
+
+        old_version = float(old_version)
+        new_version = float(new_version)
+
+        # get next major version change
+        next_major = float(math.floor(old_version + 1))
+        # get difference between versions
+        diff = round(new_version - old_version, 1)
+        if next_major != new_version and diff != 0.1:
+            return False
+        return True
 
     def get_md_key(self):
         """Create string describing each type of difference (change, addition, etc).
@@ -566,9 +554,7 @@ class DiffStix(object):
             "* Object changes: " + statusDescriptions["changes"] + "\n"
         )
         if self.minor_changes:
-            key += (
-                "* Minor object changes: " + statusDescriptions["minor_changes"] + "\n"
-            )
+            key += "* Minor object changes: " + statusDescriptions["minor_changes"] + "\n"
         if self.unchanged:
             key += "* Unchanged objects: " + statusDescriptions["unchanged"] + "\n"
         key += (
@@ -622,49 +608,36 @@ class DiffStix(object):
             childless = list(
                 filter(
                     lambda item: not self.has_subtechniques(item, True)
-                    and not (
-                        "x_mitre_is_subtechnique" in item
-                        and item["x_mitre_is_subtechnique"]
-                    ),
+                    and not ("x_mitre_is_subtechnique" in item and item["x_mitre_is_subtechnique"]),
                     items,
                 )
             )
             parents = list(
                 filter(
                     lambda item: self.has_subtechniques(item, True)
-                    and not (
-                        "x_mitre_is_subtechnique" in item
-                        and item["x_mitre_is_subtechnique"]
-                    ),
+                    and not ("x_mitre_is_subtechnique" in item and item["x_mitre_is_subtechnique"]),
                     items,
                 )
             )
             children = {
                 item["id"]: item
                 for item in filter(
-                    lambda item: ("x_mitre_is_subtechnique") in item
-                    and (item["x_mitre_is_subtechnique"]),
+                    lambda item: ("x_mitre_is_subtechnique") in item and (item["x_mitre_is_subtechnique"]),
                     items,
                 )
             }
         else:
-            childless = (
-                []
-            )  # all data sources should have data components, i.e., should have children
+            childless = []  # all data sources should have data components, i.e., should have children
             parents = list(
                 filter(
-                    lambda item: not (
-                        "x_mitre_data_source_ref" in item
-                        and item["x_mitre_data_source_ref"]
-                    ),
+                    lambda item: not ("x_mitre_data_source_ref" in item and item["x_mitre_data_source_ref"]),
                     items,
                 )
             )
             children = {
                 item["id"]: item
                 for item in filter(
-                    lambda item: ("x_mitre_data_source_ref") in item
-                    and (item["x_mitre_data_source_ref"]),
+                    lambda item: ("x_mitre_data_source_ref") in item and (item["x_mitre_data_source_ref"]),
                     items,
                 )
             }
@@ -674,35 +647,23 @@ class DiffStix(object):
         for relationship in subtechnique_of_rels:
             if relationship["target_ref"] in parentToChildren:
                 if relationship["source_ref"] in children:
-                    parentToChildren[relationship["target_ref"]].append(
-                        children[relationship["source_ref"]]
-                    )
+                    parentToChildren[relationship["target_ref"]].append(children[relationship["source_ref"]])
             else:
                 if relationship["source_ref"] in children:
-                    parentToChildren[relationship["target_ref"]] = [
-                        children[relationship["source_ref"]]
-                    ]
+                    parentToChildren[relationship["target_ref"]] = [children[relationship["source_ref"]]]
 
         for datacomponent in datacomponents:
             if datacomponent["x_mitre_data_source_ref"] in parentToChildren:
                 if datacomponent["id"] in children:
-                    parentToChildren[datacomponent["x_mitre_data_source_ref"]].append(
-                        children[datacomponent["id"]]
-                    )
+                    parentToChildren[datacomponent["x_mitre_data_source_ref"]].append(children[datacomponent["id"]])
             else:
                 if datacomponent["id"] in children:
-                    parentToChildren[datacomponent["x_mitre_data_source_ref"]] = [
-                        children[datacomponent["id"]]
-                    ]
+                    parentToChildren[datacomponent["x_mitre_data_source_ref"]] = [children[datacomponent["id"]]]
 
         # now group parents and children
         groupings = []
         for parent in childless + parents:
-            parent_children = (
-                parentToChildren.pop(parent["id"])
-                if parent["id"] in parentToChildren
-                else []
-            )
+            parent_children = parentToChildren.pop(parent["id"]) if parent["id"] in parentToChildren else []
             groupings.append(
                 {
                     "parent": parent,
@@ -751,37 +712,22 @@ class DiffStix(object):
                 """Get a section list item for the given SDO according to section type"""
                 if section == "revocations":
                     revoker = item["revoked_by"]
-                    if (
-                        "x_mitre_is_subtechnique" in revoker
-                        and revoker["x_mitre_is_subtechnique"]
-                    ):
+                    if "x_mitre_is_subtechnique" in revoker and revoker["x_mitre_is_subtechnique"]:
                         # get revoking technique's parent for display
-                        parentID = list(
-                            filter(
-                                lambda rel: rel["source_ref"] == revoker["id"],
-                                subtechnique_of_rels,
-                            )
-                        )[0]["target_ref"]
+                        parentID = list(filter(lambda rel: rel["source_ref"] == revoker["id"], subtechnique_of_rels,))[
+                            0
+                        ]["target_ref"]
                         parentName = (
-                            id_to_technique[parentID]["name"]
-                            if parentID in id_to_technique
-                            else "ERROR NO PARENT"
+                            id_to_technique[parentID]["name"] if parentID in id_to_technique else "ERROR NO PARENT"
                         )
                         return f"{item['name']} (revoked by { parentName}: [{revoker['name']}]({self.site_prefix}/{self.getUrlFromStix(revoker, True)}))"
-                    elif (
-                        "x_mitre_data_source_ref" in revoker
-                        and revoker["x_mitre_data_source_ref"]
-                    ):
+                    elif "x_mitre_data_source_ref" in revoker and revoker["x_mitre_data_source_ref"]:
                         # get revoking technique's parent for display
-                        parentID = list(
-                            filter(
-                                lambda rel: rel["id"] == revoker["id"], datacomponents
-                            )
-                        )[0]["x_mitre_data_source_ref"]
+                        parentID = list(filter(lambda rel: rel["id"] == revoker["id"], datacomponents))[0][
+                            "x_mitre_data_source_ref"
+                        ]
                         parentName = (
-                            id_to_datasource[parentID]["name"]
-                            if parentID in id_to_datasource
-                            else "ERROR NO PARENT"
+                            id_to_datasource[parentID]["name"] if parentID in id_to_datasource else "ERROR NO PARENT"
                         )
                         return f"{item['name']} (revoked by { parentName}: [{revoker['name']}]({self.site_prefix}/{self.getDataComponentUrl(id_to_datasource[parentID], item)}))"
                     else:
@@ -800,6 +746,24 @@ class DiffStix(object):
                             return f"[{item['name']}]({self.site_prefix}/{self.getDataComponentUrl(id_to_datasource[parentID], item)})"
                     return f"[{item['name']}]({self.site_prefix}/{self.getUrlFromStix(item, is_subtechnique)})"
 
+            def version(item, section):
+                if section in ["additions", "deprecations", "revocations"]:
+                    # only display current version
+                    color = (
+                        "#929393"
+                        if self.version_increment_is_valid(None, item["x_mitre_version"], section)
+                        else "#eb6635"
+                    )
+                    return f"<small style=\"color:{color}\">(v{item['x_mitre_version']})</small>"
+                else:
+                    # display previous and current version
+                    color = (
+                        "#929393"
+                        if self.version_increment_is_valid(item["previous_version"], item["x_mitre_version"], section)
+                        else "#eb6635"
+                    )
+                    return f"<small style=\"color:{color}\">(v{item['previous_version']}&#8594;v{item['x_mitre_version']})</small>"
+
             groupings = self.get_groupings(
                 obj_type=obj_type,
                 items=items,
@@ -813,28 +777,23 @@ class DiffStix(object):
             sectionString = ""
             for grouping in groupings:
                 if grouping["parentInSection"]:
-                    sectionString += f"* { placard(grouping['parent']) }\n"
+                    sectionString += f"* { placard(grouping['parent']) } { version(grouping['parent'], section) }\n"
 
-                for child in sorted(
-                    grouping["children"], key=lambda child: child["name"]
-                ):
+                for child in sorted(grouping["children"], key=lambda child: child["name"]):
                     if grouping["parentInSection"]:
-                        sectionString += f"  * {placard(child) }\n"
+                        sectionString += f"  * { placard(child) } { version(child, section) }\n"
                     else:
                         sectionString += (
-                            f"* {grouping['parent']['name']}: { placard(child) }\n"
+                            f"* {grouping['parent']['name']}: { placard(child) } { version(child, section) }\n"
                         )
 
             logger.debug(f"finished getting section list for {obj_type}/{section}")
-            # logger.debug(sectionString)
             return sectionString
 
         def getContributorSection():
             # Get contributors markdown
-            contribSection = "### Contributors to this release\n\n"
-            sorted_contributors = sorted(
-                self.release_contributors, key=lambda v: v.lower()
-            )
+            contribSection = "## Contributors to this release\n\n"
+            sorted_contributors = sorted(self.release_contributors, key=lambda v: v.lower())
 
             for contributor in sorted_contributors:
                 if contributor == "ATT&CK":
@@ -842,44 +801,35 @@ class DiffStix(object):
                 contribSection += f"* {contributor}\n"
 
             return contribSection
+
         logger.info("generating markdown string")
         content = ""
         for obj_type in self.data.keys():
             domains = ""
             for domain in self.data[obj_type]:
-                logger.debug(
-                    f"==== Generating markdown for domain: {domainToDomainLabel[domain]} --- {obj_type} ===="
-                )
-                domains += f"#### {domainToDomainLabel[domain]}\n\n"  # e.g "Enterprise"
-                # Skip mobile sections for data sources
+                logger.debug(f"==== Generating markdown for domain: {domainToDomainLabel[domain]} --- {obj_type} ====")
+                domains += f"### {domainToDomainLabel[domain]}\n\n"  # e.g "Enterprise"
+                # Skip mobile and ics sections for data sources
                 if domain == "mobile-attack" and obj_type == "datasource":
-                    logger.debug(
-                        f"Skipping - ATT&CK for Mobile does not support data sources"
-                    )
-                    domains += "ATT&CK for Mobile does not support data sources\n\n"
+                    logger.debug(f"Skipping - ATT&CK for Mobile does not support data sources")
+                    domains += f"ATT&CK for Mobile does not support data sources\n\n"
                     continue
                 domain_sections = ""
                 for section, values in self.data[obj_type][domain].items():
                     logger.debug(f"{section}: {len(values)}")
 
                     if values:  # if there are items in the section
-                        section_items = getSectionList(
-                            items=values, obj_type=obj_type, section=section
-                        )
+                        section_items = getSectionList(items=values, obj_type=obj_type, section=section)
                     else:  # no items in section
                         section_items = "* No changes\n"
 
-                    header = sectionNameToSectionHeaders[section] + ":"
+                    header = f"#### {sectionNameToSectionHeaders[section]}"
 
                     if "{obj_type}" in header:
                         if section == "additions":
-                            header = header.replace(
-                                "{obj_type}", attackTypeToTitle[obj_type]
-                            )
+                            header = header.replace("{obj_type}", attackTypeToTitle[obj_type])
                         else:
-                            header = header.replace(
-                                "{obj_type}", attackTypeToSectionName[obj_type]
-                            )
+                            header = header.replace("{obj_type}", attackTypeToSectionName[obj_type])
 
                     # e.g "added techniques:"
                     domain_sections += f"{header}\n\n{section_items}\n"
@@ -888,7 +838,7 @@ class DiffStix(object):
                 domains += f"{domain_sections}"
 
             # e.g "techniques"
-            content += f"### {attackTypeToTitle[obj_type]}\n\n{domains}"
+            content += f"## {attackTypeToTitle[obj_type]}\n\n{domains}"
 
         if self.show_key:
             key_content = self.get_md_key()
@@ -924,14 +874,10 @@ class DiffStix(object):
                 for technique in self.data["technique"][domain][status]:
                     problem_detected = False
                     if "kill_chain_phases" not in technique:
-                        logger.error(
-                            f"{technique['id']}: technique missing a tactic!! {technique['name']}"
-                        )
+                        logger.error(f"{technique['id']}: technique missing a tactic!! {technique['name']}")
                         problem_detected = True
                     if "external_references" not in technique:
-                        logger.error(
-                            f"{technique['id']}: technique missing external references!! {technique['name']}"
-                        )
+                        logger.error(f"{technique['id']}: technique missing external references!! {technique['name']}")
                         problem_detected = True
 
                     if problem_detected:
@@ -940,15 +886,11 @@ class DiffStix(object):
                     for phase in technique["kill_chain_phases"]:
                         techniques.append(
                             {
-                                "techniqueID": technique["external_references"][0][
-                                    "external_id"
-                                ],
+                                "techniqueID": technique["external_references"][0]["external_id"],
                                 "tactic": phase["phase_name"],
                                 "enabled": True,
                                 "color": statusToColor[status],
-                                "comment": status[:-1]
-                                if status != "unchanged"
-                                else status,  # trim s off end of word
+                                "comment": status[:-1] if status != "unchanged" else status,  # trim s off end of word
                             }
                         )
                         used_statuses.add(status)
@@ -1013,9 +955,7 @@ class DiffStix(object):
                 if grouping["parentInSection"]:
                     new_values.append(grouping["parent"])
 
-                for child in sorted(
-                    grouping["children"], key=lambda child: child["name"]
-                ):
+                for child in sorted(grouping["children"], key=lambda child: child["name"]):
                     new_values.append(child)
 
             return new_values
@@ -1026,16 +966,12 @@ class DiffStix(object):
 
         for obj_type, domains in self.data.items():
             for domain, sections in domains.items():
-                logger.debug(
-                    f"===== Generating domain: {domainToDomainLabel[domain]} --- {obj_type} ====="
-                )
+                logger.debug(f"===== Generating domain: {domainToDomainLabel[domain]} --- {obj_type} =====")
                 changes_dict[domain][obj_type] = {}
 
                 for section, values in sections.items():
                     # new_values includes parents & children mixed (e.g. techniques/sub-techniques, data sources/components)
-                    new_values = cleanup_values(
-                        items=values, obj_type=obj_type, section=section
-                    )
+                    new_values = cleanup_values(items=values, obj_type=obj_type, section=section)
                     changes_dict[domain][obj_type][section] = new_values
 
         # always add contributors
@@ -1062,9 +998,7 @@ def markdown_to_index_html(markdown_outfile, content):
     # Center content
     html_string = """<div style='max-width: 55em;margin: auto;margin-top:20px;font-family: "Roboto", sans-serif;'>"""
     html_string += "<meta charset='utf-8'>"
-    html_string += (
-        "<h1 style='text-align:center;'>Changes between ATT&CK STIX bundles</h1>"
-    )
+    html_string += "<h1 style='text-align:center;'>Changes between ATT&CK STIX bundles</h1>"
     html_string += markdown.markdown(content)
     html_string += "</div>"
 
@@ -1073,6 +1007,7 @@ def markdown_to_index_html(markdown_outfile, content):
     outfile.close()
 
     logger.info("finished writing HTML to file")
+
 
 def layers_dict_to_files(outfiles, layers):
     """Print the layers dict passed in to layer files."""
@@ -1129,8 +1064,8 @@ def get_parsed_args():
         type=str,
         nargs="+",
         metavar=("OBJ_TYPE", "OBJ_TYPE"),
-        choices=["technique", "software", "group", "mitigation", "datasource"],
-        default=["technique", "software", "group", "mitigation", "datasource"],
+        choices=["technique", "software", "group", "campaign", "mitigation", "datasource"],
+        default=["technique", "software", "group", "campaign", "mitigation", "datasource"],
         help="which types of objects to report on. Choices (and defaults) are %(choices)s",
     )
     parser.add_argument(
@@ -1245,9 +1180,7 @@ def get_parsed_args():
         parser.error("--use-mitre-cti and -old cannot be used together")
 
     if not args.markdown and args.layers is None:
-        logger.error(
-            "Script doesn't output anything unless -markdown and/or -layers are specified."
-        )
+        logger.error("Script doesn't output anything unless -markdown and/or -layers are specified.")
         logger.error("Run 'python3 diff_stix.py -h' for usage instructions")
         exit()
 
@@ -1256,9 +1189,7 @@ def get_parsed_args():
 
     if args.layers is not None:
         if len(args.layers) not in [0, 3]:
-            parser.error(
-                "-layers requires exactly three files to be specified or none at all"
-            )
+            parser.error("-layers requires exactly three files to be specified or none at all")
 
     return args
 
@@ -1274,7 +1205,7 @@ def get_new_changelog_md(
     old: str = None,
     show_key: bool = False,
     site_prefix: str = "",
-    types: List[str] = ["technique", "software", "group", "mitigation", "datasource"],
+    types: List[str] = ["technique", "software", "group", "campaign", "mitigation", "datasource"],
     use_taxii: bool = False,
     use_mitre_cti: bool = False,
     verbose: bool = False,